match/serializers.py
from rest_framework import serializers
from .models import Match

class MatchSerializer(serializers.ModelSerializer):
    class Meta:
        model = Match
        fields = ('id', 'user_from', 'user_to', 'liked', 'timestamp')
